+++ markket/api.ts
@@ -3,7 +3,7 @@ import { Store } from './store.d';
 import { Page } from './page.d';
 import qs from 'qs';
 
-export const MARKKET_URL = process.env.NEXT_PUBLIC_MARKKET_API || 'https://de.markket.place/';
+export { markketClient, MARKKET_URL } from './api.markket';
 
 export type { StrapiResponse, FetchOptions };
 
@@ -22,71 +22,7 @@ interface EnhancedFetchOptions extends Omit<FetchOptions, 'filters'> {
   };
 };
 
-type fetchOptions = {
-  method?: string;
-  headers?: any;
-  body?: any;
-};
 
-/**
- * Utitilities to easily communitcate with our routes in /api/markket
- */
-export class markketClient {
-  private baseUrl: string;
-  private token: string;
-
-  constructor() {
-    if (typeof window == 'undefined') {
-      this.baseUrl = MARKKET_URL as string;
-      this.token = '';
-    } else {
-      this.baseUrl = window.location.origin;
-      this.token = '';
-    }
-  };
-
-  public readToken = () => {
-    if (typeof window == 'undefined') { return null; }
-
-    const _string = localStorage.getItem('markket.auth');
-    const _json = _string ? JSON.parse(_string) : {};
-    const { jwt } = _json;
-
-    this.token = jwt;
-    return jwt;
-  };
-
-  public fetch = async (url: string, options: fetchOptions) => {
-    const token = this.readToken();
-
-    console.log({ url });
-    const _url = new URL(url, this.baseUrl);
-
-    const response = await fetch(_url.toString(), {
-      ...options,
-      method: options?.method || 'GET',
-      headers: {
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json',
-        ...(options.headers || {}),
-      },
-    });
-
-    return await response.json();
-  };
-
-  public post = async (url: string, options: any) => {
-    return this.fetch(url, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        ...(options.headers || {}),
-      },
-      body: JSON.stringify(options?.body || {}),
-    });
-  };
-
-};
 
 export class StrapiClient {
   private baseUrl: string;
